# 15-spring-boot-rest-api-unit-test-app/src/main/java/com/tekleads/domain/Contacts.java
package com.tekleads.domain;

import java.util.List;

import javax.xml.bind.annotation.XmlRootElement;

@XmlRootElement(name="contacts")
public class Contacts {
	
	private List<Contact> contacts;

	public List<Contact> getContacts() {
		return contacts;
	}

	public void setContacts(List<Contact> contacts) {
		this.contacts = contacts;
	}
}