# 15-spring-boot-rest-api-unit-test-app/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "Contacts Operations",
		"version": "1.1",
		"title": "Contact-Api",
		"termsOfService": "www.tekleads.com",
		"contact": {
			"name": "TEK Leads",
			"url": "www.tekleads.com",
			"email": "tekleads4u@gmail.com"
		},
		"license": {
			"name": "www.apache.org/licence",
			"url": "www.apache.org"
		}
	},
	"host": "localhost:9091",
	"basePath": "/",
	"tags": [
		{
			"name": "contact-rest-resource",
			"description": "Contact Rest Resource"
		},
		{
			"name": "welcome-rest-resource",
			"description": "Welcome Rest Resource"
		}
	],
	"paths": {
		"/addContact": {
			"post": {
				"tags": [
					"contact-rest-resource"
				],
				"summary": "addContact",
				"operationId": "addContactUsingPOST",
				"consumes": [
					"application/xml",
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "c",
						"description": "c",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Contact"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/addContacts": {
			"post": {
				"tags": [
					"contact-rest-resource"
				],
				"summary": "addContacts",
				"operationId": "addContactsUsingPOST",
				"consumes": [
					"application/xml",
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "contacts",
						"description": "contacts",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Contacts"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/contacts": {
			"get": {
				"tags": [
					"contact-rest-resource"
				],
				"summary": "getAllContacts",
				"operationId": "getAllContactsUsingGET",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/xml",
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Contacts"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/contacts/{cid}": {
			"get": {
				"tags": [
					"contact-rest-resource"
				],
				"summary": "getContactById",
				"operationId": "getContactByIdUsingGET",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/xml",
					"application/json"
				],
				"parameters": [
					{
						"name": "cid",
						"in": "path",
						"description": "cid",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Contact"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/header": {
			"get": {
				"tags": [
					"contact-rest-resource"
				],
				"summary": "readHeaders",
				"operationId": "readHeadersUsingGET",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "secret-key",
						"in": "header",
						"description": "secret-key",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/rest/greet/{name}": {
			"get": {
				"tags": [
					"welcome-rest-resource"
				],
				"summary": "greetMsg",
				"operationId": "greetMsgUsingGET",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "name",
						"in": "path",
						"description": "name",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/rest/welcome": {
			"get": {
				"tags": [
					"welcome-rest-resource"
				],
				"summary": "welcomeMsg",
				"operationId": "welcomeMsgUsingGET",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "name",
						"in": "query",
						"description": "name",
						"required": false,
						"type": "string",
						"default": "John"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		}
	},
	"definitions": {
		"Contacts": {
			"type": "object",
			"properties": {
				"contacts": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Contact"
					}
				}
			}
		},
		"Contact": {
			"type": "object",
			"properties": {
				"contactId": {
					"type": "integer",
					"format": "int32"
				},
				"contactName": {
					"type": "string"
				},
				"contactNumber": {
					"type": "integer",
					"format": "int64"
				},
				"gender": {
					"type": "string"
				}
			}
		}
	}
}